backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Database
sqlalchemy==2.0.23
neo4j==5.13.0

# LLM Integration
openai==1.3.0
google-generativeai==0.3.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Text Processing
scikit-learn==1.3.2
jieba==0.42.1
nltk==3.8.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0